requirements.txt
telethon
python-dotenv
loguru
psycopg2-binary
dbt-postgres
ultralytics
fastapi
uvicorn
httpx
dagster
dagster-webserver
pytest
flake8
pandas
matplotlib
nest_asyncio